requirements.txt

Flask
python-dotenv
PyPDF2
openai>=1.0.0
